修改表结构设计的不合理的地方

+++ gulimall-model/gulimall-admin-model/src/main/java/cn/gulimall/demo/model/po/SysRoleMenu.java
@@ -3,19 +3,19 @@ package cn.gulimall.demo.model.po;
 import com.mybatisflex.annotation.Id;
 import com.mybatisflex.annotation.KeyType;
 import com.mybatisflex.annotation.Table;
-import java.io.Serializable;
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
 import java.io.Serial;
+import java.io.Serializable;
 
 /**
  * 角色与菜单对应关系 实体类。
  *
  * @author Z
- * @since 2024-05-18
+ * @since 2024-07-26
  */
 @Data
 @Builder
@@ -28,16 +28,16 @@ public class SysRoleMenu implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @Id(keyType = KeyType.Auto)
-    private Long id;
+    private Integer id;
 
     /**
      * 角色ID
      */
-    private Long roleId;
+    private Integer roleId;
 
     /**
      * 菜单ID
      */
-    private Long menuId;
+    private Integer menuId;
 
 }
