+++ gulimall-model/gulimall-admin-model/src/main/java/cn/gulimall/demo/model/po/SysUserRole.java
@@ -3,19 +3,19 @@ package cn.gulimall.demo.model.po;
 import com.mybatisflex.annotation.Id;
 import com.mybatisflex.annotation.KeyType;
 import com.mybatisflex.annotation.Table;
-import java.io.Serializable;
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
 import java.io.Serial;
+import java.io.Serializable;
 
 /**
  * 用户与角色对应关系 实体类。
  *
  * @author Z
- * @since 2024-05-18
+ * @since 2024-07-26
  */
 @Data
 @Builder
@@ -28,16 +28,16 @@ public class SysUserRole implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @Id(keyType = KeyType.Auto)
-    private Long id;
+    private Integer id;
 
     /**
      * 用户ID
      */
-    private Long userId;
+    private Integer userId;
 
     /**
      * 角色ID
      */
-    private Long roleId;
+    private Integer roleId;
 
 }
